Move integration testing helper regtest module to sbtc crate
