+++ sbtc/src/testing/mod.rs
@@ -1,0 +1,2 @@
+//! All testing utility functions.
+pub mod regtest;
