+++ signer/tests/bitcoin/bitcoin.conf
@@ -626,7 +626,7 @@ chain=regtest
 # then connects normally using the
 # rpcuser=<USERNAME>/rpcpassword=<PASSWORD> pair of arguments. This
 # option can be specified multiple times
-rpcauth=alice:5d16a7f9a0caa700499c8025fa32400d$305bf76ef314d4669912c21db105bd14447f4066a822d8b90bbcd0d29a8bc599
+rpcauth=devnet:05bb3a3c6814d5980f0b43e2ae21c8a6$4c84d1fbda611224f9209201dd119681eaafe2ccf62935d028dd956ff80f584e
 
 # Bind to given address to listen for JSON-RPC connections. Do not expose
 # the RPC server to untrusted networks such as the public internet!
@@ -641,7 +641,7 @@ rpcauth=alice:5d16a7f9a0caa700499c8025fa32400d$305bf76ef314d4669912c21db105bd144
 #rpccookiefile=<loc>
 
 # Password for JSON-RPC connections
-rpcpassword=pw
+rpcpassword=devnet
 
 # Listen for JSON-RPC connections on <port> (default: 8332, testnet:
 # 18332, signet: 38332, regtest: 18443)
@@ -651,7 +651,7 @@ rpcpassword=pw
 #rpcthreads=<n>
 
 # Username for JSON-RPC connections
-rpcuser=alice
+rpcuser=devnet
 
 # Set a whitelist to filter incoming RPC calls for a specific user. The
 # field <whitelist> comes in the format: <USERNAME>:<rpc 1>,<rpc
