+++ signer/tests/integration/main.rs
@@ -2,6 +2,5 @@ mod block_notifier;
 mod contracts;
 mod postgres;
 mod rbf;
-pub mod regtest;
 mod transaction_signer;
 mod utxo_construction;
